+++ src/models/inFile/InFileOrder.ts
@@ -1,10 +1,11 @@
 import {injectable} from 'inversify';
-import {readFromFile, writeToFile} from 'src/models/inFile/dbHelpers';
 import {Optional} from 'src/utils/typeUtils';
 import {v4 as uuid} from 'uuid';
 import {Order, OrderFilterType, OrderType} from '../interfaces';
+import {helper} from './dbHelper';
 
 const DB_FILE_NAME = 'Order.json';
+const h = helper(DB_FILE_NAME);
 
 @injectable()
 export class InFileOrder implements Order {
@@ -13,18 +14,18 @@ export class InFileOrder implements Order {
     constructor() {
         this.orders = [];
 
-        readFromFile(DB_FILE_NAME).then(res => {
+        h.readFromFile().then(res => {
             this.orders = res as OrderType[];
         }).catch(() => {
             this.orders = [];
-            writeToFile(DB_FILE_NAME, JSON.stringify(this.orders));
+            h.writeToFile(this.orders);
         });
     }
 
     async save(order: Optional<OrderType, 'id'>): Promise<OrderType> {
         const dbOrder = {...order, id: uuid()};
         this.orders.push(dbOrder);
-        writeToFile(DB_FILE_NAME, JSON.stringify(this.orders));
+        h.writeToFile(this.orders);
 
         return dbOrder;
     }
